fix initial build and improve shutdown script
gen_init_cpio has to be build before mkinitramfs.sh works
shutdown should remount readonly in case of failure. lazy-unmounting is likely to not unmount anything at all because the machine will be off by the time the filesystem would be unused
also use /boot/linux as uImage filename. it is more intuitive