add card removal support to smartcard GPIO
when the card is removed, or on short circuits, the interface IC shuts down VCC to the card.
make linux react to that and pull all pins low. this has two functions:
a) it allows userspace to notice that by reading the value of the power control pin, which goes low
b) the interface IC will not allow pulling pins high again when the card is reinserted, unless they have been pulled low before