implement boot-time default underscan, and make framebuffer cacheable
the framebuffer now defaults to 640x512 for PAL and 640x432 for NTSC.
these values roughly match the borders that are seen on boot when the SCART switch hasn't been shut down properly.

also map the framebuffer cacheable. speeds things up enormously, to the point where moving a window around in X11 is almost as fast as the mouse.
curiously, there is no noticeable screen corruption due to unflushed caches. (admittedly, the cursor doesn't blink very reliably. but that's probably tolerable.)