implement proper framebuffer blanking, and shutdown unnecessary secondary DACs
actually blank the screen when asked to. otherwise it corrupts on next update, which is extremely confusing.
but disable automatic screensaver blanking in initramfs. it's probably connected to a TV and blanking a TV is annoying.

also shuts down the secondary video channel of the PNX8510, as well as its audio part. these are unused and the chip runs very hot anyway.