document clock module and improve power-downs
document what each PLL and DDS unit is connected to, and some of the clock control registers in the clock module.
use these symbolic names in framebuffer driver and document what each block of the init code does, where known.

also slow down the clock to unused modules (TM32 1/2 and QVCP2), but don't stop them completely. stopping them completely hangs the system when the module is accessed.
set the relevant PLL to 1.6875MHz instead. compared to 270MHz for the TM32s and 27MHz for the QVCP2, this should still save some power.