restore address space and high address enable after reading presence and interrupt status
when the module is extracted, the chip resets the address space to attribute memory, and clears the high address enable bit.
restore them after reading the chip's interrupt or presence status, so that userspace doesn't have to manually restore them.
unfortunately, this slows down reading (and polling) the status files. for detection, there's not much we can do.
for the interrupt status, we can optimize that to only restore status when there is an actual module movement being reported.