implement interrupt reporting in /dev/cimax-interrupt
when userspace read from the /dev/cimax-interrupt device, the kernel either reports the pending interrupt, or blocks until an interrupt happens.
the reported interrupts are "presence" (higher priority) for module detection, and "irq" (lower priority) for an IRQ from the module.
as a useful side effect, the interrupt handler also restores the selected address space and enable the high address bits, so userspace doesn't have to worry asbout that.