implement zero-copy audio driver
make the hardware buffer uncached-mmapped right into userspace. this saves some memcpy overhead which is largely imaginary, but saves memory and is cleaner.
because the buffer is statically allocated, it depends on the client asking for exactly the right buffer size. this doesn't work for full-page buffers, so reduce their size a bit.