use the frontpanel display chip the way it is meant to
this means having STROBE and CLOCK idle high, and using STROBE as a chip select (essentially),
as well as setting the number of digits to be displayed on startup (else the LEDs might not be scanned).

also use true linear brightness values in the range 0..15, with 0=blanked.
the chip doesn't support all of them, but it's possible to just set the nearest one instead.