always keep output clock running for both QVCPs
for QVCP 2, this is required else accessing the DACs hangs the system. for
QVCP 1, the Anabel fails to respond on I²C, but doesn't hang. unfortunately,
it needs more than the ~1.6MHz shutdown clock from the PLLs, so just keep it
running the full at 27MHz. doesn't matter much for the power consumption,
fortunately.

also wait a bit before trying to shutdown QVCP 2 after turining on the
clocks. if they have been stopped before for some reason, this could hang
entire the system.