cimax: report new status on presence change
this way, userspace doesn't have to read the sysfs file to determine what
actually happened, avoiding a slight race condition where the status changes
again before the sysfs file is read.

the new code will always report the status of the card as it was at some
point between the interrupt and when read() returns. this may be the same
as the old status if the card has be ejected and reinserted; the fact that
an interrupt was reported still means that it might be a different card now.
also note that the card may already have been ejected again when read()
returns; there is nothing that can be done against that except telling users
not to yank out cards.