present entire cimax address space as a single linear range
this is somewhat incorrect because A16/A17 actually come out as A24/A25 on
the connector. however, the intended use is as an 8-bit ISA-like port, with
the address lines correctly permuted between the plug-in card and whatever
board it is connected to. for this use, the simpler the software interface
the better, so presenting a single memory block is more convenient.