framebuffer: underscan fake HD by 5%
TVs are supposed to just show the picture for HD signals, but some overscan
it anyway. at least they don't overscan as much as old SD TVs, so 5%
underscan should be ok.

the alternative would be to use a pixel clock of 51.975MHz and increase
screen width to 1848 pixels, ie. just squeezing the pixels more tightly.
this works for horizontal scaling, but would lead to a very odd vertical
scaling factor that cannot be filtered easily.