simplify interrupt setup
instead of always restoring priority from a table, just set it to 1 during
init and never touch it again. reduces the number of unnecessary memory
accesses and works as well as an effectively all-ones table. (the only non-
one value was the interrupt for OHCI, which is unused.)

instead of initializing interrupts as active high / active low according to
a table, always initialize them as active high and allow setting them to
active low later. as long as they are disabled, no spurious interrupts will
be generated.