initramfs: mount initrd root readonly; next-method bootfail
mounting the root readonly makes it harder to write to it accidentally.
it will still replay the journal for journalled filesystems, but it will
do so using the same code that would be used later anyway, so that's ok.

bootfail now uses the next fields as new bootfail field, so eg.
  bootfail=UUID=xyz:/boot/fail:reboot
tries to load /boot/fail, and reboots if that fails.

make check for root usability slightly more robust: /root actually has to
be a mountpoint, else switch_root fails.