fix extremely rare exception mis-routing
very rarely, exceptions can end up in the wrong handler. in each case
observed, the wrong exception is taken, but the value in CP0_CAUSE clearly
identifies it as a TLB refill (exception 2). for example, it can end up
being reported as "reserved 2", or as a "machine check" with a cause of
0x90800008. so for, this has only been observed for GCC 4.9.

looks like there is a hazard from an exception occuring to the instruction
reading CP0_CAUSE. this is normally the first instruction in the exception
handler (as designed), but apparently on PR4450 it isn't reliable until the
second instruction. a simple NOP in front of things slows down every single
exception, but appears to get rid of all exception weirdness.

there is almost certainly a matching hazard in the TLB refill handler.
fortunately, the first instruction in that exception handler is a lui, not
an mfc0...